翻译修改，以及uid修改

+++ Android/data/src/main/java/com/agora/data/provider/IRoomProxy.java
@@ -16,7 +16,7 @@ public interface IRoomProxy {
     Member getMemberById(@NonNull String memberId);
 
     @Nullable
-    Member getMemberByStramId(long streamId);
+    Member getMemberByStramId(int streamId);
 
     Maybe<Member> getMember(@NonNull String roomId, @NonNull String userId);
